Match BattleIntroOpponent1SendsOutMonAnimation
